Use more common build variables, fix build rules
See
https://www.gnu.org/software/make/manual/html_node/Implicit-Variables.html
for reference (I know that manual is GNU-specific, but most of this is
not GNU-specific).

Since the code is C++, appropriate compiler and flags variables are
CXX and CXXFLAGS. Pre-processor directives go in CPPFLAGS. Libraries
are specified in LDLIBS; the standard C++ library is included
automatically when appropriate and does not need to be specified.

The target directory should be created (for use with non-default
DESTDIR). To allow this, specify 0755 instead of 0555.

Automatic variables are used where appropriate.

Signed-off-by: Stephen Kitt <steve@sk2.org>